Move world clock widget constants and helper methods
